Feat/packages update

+++ src/controllers/todos.ts
@@ -1,13 +1,10 @@
 import { Request, Response } from "express";
-import { getCustomRepository } from "typeorm";
-import {
-  TodosRepository,
-  UpdateTodoRequestBody,
-} from "src/db/repository/TodoRepository";
-import { StatusCodes } from "http-status-codes";
-import { UsersRepository } from "src/db/repository/UserRepository";
 import omit from "lodash/omit";
-import { Todo } from "src/db/entity/Todo";
+import { Todo } from "@prisma/client";
+import { type UpdateTodoRequestBody } from "src/db/repository/TodoRepository";
+import * as todosRepository from "src/db/repository/TodoRepository";
+import * as userRepository from "src/db/repository/UserRepository";
+import { StatusCodes } from "http-status-codes";
 
 const formatTodo = (todo: Todo) => ({
   ...omit(todo, "id"),
@@ -19,23 +16,19 @@ export async function createTodo(req: Request, res: Response) {
   const { userId } = req.params;
   const { description }: { description: string } = req.body;
 
-  const todosRepository = getCustomRepository(TodosRepository);
-  const userRepositor = getCustomRepository(UsersRepository);
-
   try {
-    const user = await userRepositor.getUserById(Number(userId));
+    const newTodo = await todosRepository.addTodo({
+      userId: Number(userId),
+      description,
+    });
 
-    const newTodo = await todosRepository.addTodo(user, description);
-
-    const {
-      raw: [{ id }],
-    } = newTodo;
-
-    const todo = await todosRepository.getTodoById(id);
+    const todo = await todosRepository.getTodoById(newTodo.id);
 
     res.status(StatusCodes.CREATED);
 
     res.send(formatTodo(todo));
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
   } catch (err) {
     res.sendStatus(StatusCodes.INTERNAL_SERVER_ERROR);
   }
@@ -44,13 +37,16 @@ export async function createTodo(req: Request, res: Response) {
 export async function getTodo(req: Request, res: Response) {
   const { uuid, userId } = req.params;
 
-  const todosRepository = getCustomRepository(TodosRepository);
-  const userRepositor = getCustomRepository(UsersRepository);
+  const user = await userRepository.getUserById(Number(userId));
 
-  const user = await userRepositor.getUserById(Number(userId));
-  const todo = await todosRepository.getTodoByUuid(user, uuid);
+  if (!user) {
+    res.sendStatus(StatusCodes.NOT_FOUND);
+    return;
+  }
 
-  if (todo === undefined) {
+  const todo = await todosRepository.getTodoByUuid({ userId: user.id, uuid });
+
+  if (!todo) {
     res.sendStatus(StatusCodes.NOT_FOUND);
   } else {
     res.send(formatTodo(todo));
@@ -60,17 +56,14 @@ export async function getTodo(req: Request, res: Response) {
 export async function getTodos(req: Request, res: Response) {
   const { userId } = req.params;
 
-  const todosRepository = getCustomRepository(TodosRepository);
-  const userRepositor = getCustomRepository(UsersRepository);
+  const user = await userRepository.getUserById(Number(userId));
 
-  const user = await userRepositor.getUserById(Number(userId));
-
-  if (user === undefined) {
+  if (!user) {
     res.sendStatus(StatusCodes.NOT_FOUND);
     return;
   }
 
-  const todos = await todosRepository.getUsersTodos(user);
+  const todos = await todosRepository.getUsersTodos(user.id);
 
   res.send({
     todos: todos.map((todo) => formatTodo(todo)),
@@ -80,18 +73,22 @@ export async function getTodos(req: Request, res: Response) {
 export async function deleteTodo(req: Request, res: Response) {
   const { uuid, userId } = req.params;
 
-  const todosRepository = getCustomRepository(TodosRepository);
-  const userRepositor = getCustomRepository(UsersRepository);
+  const user = await userRepository.getUserById(Number(userId));
 
-  const user = await userRepositor.getUserById(Number(userId));
-  const todo = await todosRepository.getTodoByUuid(user, uuid);
+  if (!user) {
+    res.sendStatus(StatusCodes.NOT_FOUND);
+    return;
+  }
+  
+  const todo = await todosRepository.getTodoByUuid({ userId: user.id, uuid });
 
-  if (todo === undefined) {
+
+  if (!todo) {
     res.sendStatus(StatusCodes.NOT_FOUND);
     return;
   }
 
-  await todosRepository.deleteTodo(user, uuid);
+  await todosRepository.deleteTodo({ userId: user.id, uuid });
 
   res.sendStatus(StatusCodes.NO_CONTENT);
 }
@@ -101,23 +98,33 @@ export async function updateTodo(req: Request, res: Response) {
 
   const { description, is_done }: UpdateTodoRequestBody = req.body;
 
-  const todosRepository = getCustomRepository(TodosRepository);
-  const userRepositor = getCustomRepository(UsersRepository);
+  const user = await userRepository.getUserById(Number(userId));
 
-  const user = await userRepositor.getUserById(Number(userId));
-  const todo = await todosRepository.getTodoByUuid(user, uuid);
-
-  if (todo === undefined) {
+  if (!user) {
     res.sendStatus(StatusCodes.NOT_FOUND);
     return;
   }
 
-  await todosRepository.updateTodo(user, uuid, {
-    description,
-    is_done,
+  const todo = await todosRepository.getTodoByUuid({ userId: user.id, uuid });
+
+  if (!todo) {
+    res.sendStatus(StatusCodes.NOT_FOUND);
+    return;
+  }
+
+  await todosRepository.updateTodo({
+    userId: user.id,
+    uuid,
+    body: {
+      description,
+      is_done,
+    },
   });
 
-  const updatedTodo = await todosRepository.getTodoByUuid(user, uuid);
+  const updatedTodo = await todosRepository.getTodoByUuid({
+    userId: user.id,
+    uuid,
+  });
 
   res.send(formatTodo(updatedTodo));
 }
